requirements.txt
# Telegram
telethon>=1.34.0
python-telegram-bot>=21.0

# AI/ML
anthropic>=0.40.0
voyageai>=0.3.0
chromadb>=0.5.0

# Database
aiosqlite>=0.19.0

# Scheduling
apscheduler>=3.10.0
pytz>=2024.1

# Web Dashboard
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.0

# Config & Utils
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Logging
structlog>=24.0.0

# HTTP
httpx>=0.27.0
